# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: Accounts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Categories
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Products
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Suppliers
...
  /credits:
    get:
      operationId: listCredits
      tags: [credits]
      summary: List all credit records
      responses:
        "200":
          description: Credits list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Credit"
    post:
      operationId: createCredit
      tags: [credits]
      summary: Create a credit record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credit"
  /credits/{id}/pay:
    post:
      operationId: payCredit
      tags: [credits]
      summary: Record a payment against a credit
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayCreditRequest"
      responses:
        "200":
          description: Updated credit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credit"

  # Wallets
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Auth
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    # User
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
        locationId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, username, name, role, isActive, createdAt]
...
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        balance:
          type: string
        currency:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, type, balance, currency, isActive, createdAt]
    CreateAccountRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        balance:
          type: string
        currency:
          type: string
      required: [name, type, balance, currency]
    UpdateAccountRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        isActive:
          type: boolean

    # Category
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, type, createdAt]
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: ["string", "null"]
      required: [name, type]
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]

    # Product
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        creditBalance:
          type: string
        createdAt:
          type: string
      required: [id, name, creditBalance, createdAt]
    CreateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required: [name]
    UpdateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]

    # Supplier
...
    Credit:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        partyId:
          type: integer
        partyName:
          type: string
        partyType:
          type: string
        amount:
          type: string
        paidAmount:
          type: string
        remainingAmount:
          type: string
        dueDate:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        userId:
          type: integer
        createdAt:
          type: string
      required: [id, type, partyId, partyName, partyType, amount, paidAmount, remainingAmount, status, userId, createdAt]
    CreateCreditRequest:
      type: object
      properties:
        type:
          type: string
        partyId:
          type: integer
        partyType:
          type: string
        amount:
          type: string
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        userId:
          type: integer
      required: [type, partyId, partyType, amount, userId]
    PayCreditRequest:
      type: object
      properties:
        payAmount:
          type: string
        accountId:
          type: ["integer", "null"]
      required: [payAmount]

    # Wallet
...